move TCountersPrinter to 'cloud/storage' to be able to use it in filestore
