+++ cloud/blockstore/libs/storage/core/histogram_ut.cpp
@@ -1,6 +1,6 @@
 #include "histogram.h"
 
-#include <cloud/blockstore/libs/storage/testlib/counters_printer.h>
+#include <cloud/storage/core/libs/diagnostics/counters_printer.h>
 #include <cloud/storage/core/libs/diagnostics/histogram_types.h>
 
 #include <library/cpp/testing/unittest/registar.h>
