+++ cloud/blockstore/libs/storage/testlib/ya.make
@@ -1,7 +1,6 @@
 LIBRARY(blockstore-libs-storage-testlib)
 
 SRCS(
-    counters_printer.cpp
     diagnostics.cpp
     disk_agent_mock.cpp
     disk_registry_proxy_mock.cpp
